canto_project_official_yue_bridge_demo_v2/requirements.txt
streamlit==1.40.1
pillow>=10.0
torch>=2.1
torchvision>=0.16
transformers==4.49.0
peft>=0.10.0
accelerate>=1.2.1
huggingface_hub>=0.24
soundfile>=0.12
sentencepiece>=0.2.0
einops>=0.8.0
matplotlib>=3.7
# RAG dependencies (used when "Use RAG" is enabled for InternVL2-4B)
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
